PIL-2576 update subscription cache after amendments (#131)
Currently, when users reach the end of the "Manage contact details"
journey, we submit the amendment to ETMP, and when it is successful, we
drop the record from the cache. This is usually fine, because the user
is directed to the homepage, which always populates the cache on load.

However, if the user double-clicks the submit amendment button on the
CYA page, the user sees an error in chromium browsers:
- The frontend fetches the subscription from the backend's cache before
processing any submissions.
- The first click proceeds through the FE, to the BE, and the
subscription is dropped from the cache.
- Before the browser renders the result of this, the second click
attempts to read the sub from cache before processing, which fails
because the sub doesn't exist in cache.

This PR changes the behaviour on amend to avoid dropping the cache, and
instead fetches up-to-date state from ETMP after the amendment has
succeeded in a read-after-write pattern. This avoids the empty cache
situation on double-click.